[refactor] 공통 레이아웃 컴포넌트 디렉토리 이동
