+++ src/components/common/Container/index.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-export const Container = styled.section`
+const Container = styled.section`
   width: 100%;
   display: flex;
   flex-direction: column;
@@ -19,3 +19,5 @@ export const Container = styled.section`
     padding: 0 0.5rem;
   }
 `;
+
+export default Container;
