+++ src/components/common/Layout/index.tsx
@@ -1,8 +1,9 @@
 import { ReactNode } from 'react';
 import styled from 'styled-components';
 
+import DesktopCard from './DesktopCard';
+
 import GNB from '@/components/common/GNB';
-import DesktopPage from '@/pages/DesktopPage';
 
 interface LayoutProps {
   router: ReactNode;
@@ -11,7 +12,7 @@ interface LayoutProps {
 const Layout = ({ router }: LayoutProps) => {
   return (
     <MainContainer>
-      <DesktopPage />
+      <DesktopCard />
       <PageContainer>
         {router}
         <GNB />
